[REFACTOR] NumberElementCollection 객체를 재사용하기로 결정하고 리펙토링, 모든 테스트 통과 확인
