+++ src/main/java/lotto/LottoTicket.java
@@ -3,6 +3,7 @@ package lotto;
 import lotto.exception.DuplicateNumberException;
 import lotto.exception.OutOfCountException;
 import lotto.exception.OutOfMaxNumberException;
+import model.NumberElementCollection;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,16 +14,13 @@ public class LottoTicket {
 
 	private static final int MAX_COUNT = 6;
 
-	List<Integer> numbers = new ArrayList<>();
+	NumberElementCollection numbers = new NumberElementCollection(MAX_COUNT);
 
 	public void add(int number) {
 		if(numbers.contains(number)){
 			throw new DuplicateNumberException();
 		}
 
-		if(numbers.size() == MAX_COUNT){
-			throw new OutOfCountException();
-		}
 
 		if(number > MAX_NUMBER){
 			throw new OutOfMaxNumberException();
