+++ src/main/java/model/NumberElement.java
@@ -1,7 +1,6 @@
-package calculator;
+package model;
 
 import java.util.Objects;
-import java.util.Optional;
 
 public class NumberElement {
