+++ src/test/java/calculator/NumberElementCollectionTest.java
@@ -1,6 +1,7 @@
 package calculator;
 
-import org.junit.jupiter.api.Test;
+import model.NumberElement;
+import model.NumberElementCollection;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -8,7 +9,6 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class NumberElementCollectionTest {
