+++ src/test/java/lotto/LottoTicketTest.java
@@ -6,7 +6,6 @@ import lotto.exception.OutOfMaxNumberException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 class LottoTicketTest {
